refactor: extract RoomsPage component and consolidate create_room actions
