+++ src/app/_components/RoomsPage.jsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
-import { createClient } from '@/utils/supabase/client';
+import { getUserRooms } from '../actions';
 import Box from '@mui/joy/Box';
 import CircularProgress from '@mui/joy/CircularProgress';
 import Button from '@mui/joy/Button';
@@ -20,17 +20,7 @@ export default function RoomsPage() {
 
   useEffect(() => {
     const loadRooms = async () => {
-      const supabase = createClient();
-      const { data: { session }, error } = await supabase.auth.getSession();
-
-      if (error || !session) {
-        router.push('/login');
-        return;
-      }
-
-      setFirstName(session.user.user_metadata?.full_name?.split(' ')[0] || 'there');
-
-      const cachedRaw = localStorage.getItem(ROOMS_CACHE_KEY);
+      const cachedRaw = typeof localStorage !== 'undefined' ? localStorage.getItem(ROOMS_CACHE_KEY) : null;
       const cached = cachedRaw ? JSON.parse(cachedRaw) : null;
 
       if (!navigator.onLine) {
@@ -40,23 +30,8 @@ export default function RoomsPage() {
       }
 
       try {
-        const { data: userRecord } = await supabase
-          .from('Users')
-          .select('id')
-          .eq('email', session.user.email)
-          .single();
-
-        const { data: memberships } = await supabase
-          .from('UserRooms')
-          .select('room_id, role, Rooms(id, admin, members)')
-          .eq('user_id', userRecord.id);
-
-        const roomList = (memberships || []).map(m => ({
-          id: m.room_id,
-          role: m.role,
-          ...m.Rooms,
-        }));
-
+        const { rooms: roomList, firstName: name } = await getUserRooms();
+        setFirstName(name);
         localStorage.setItem(ROOMS_CACHE_KEY, JSON.stringify(roomList));
         setRooms(roomList);
       } catch {
@@ -66,7 +41,7 @@ export default function RoomsPage() {
     };
 
     loadRooms();
-  }, [router]);
+  }, []);
 
   if (loading || rooms === null) {
     return (
@@ -78,62 +53,62 @@ export default function RoomsPage() {
 
   return (
     <>
-    <Box sx={{ maxWidth: 480, mx: 'auto', mt: 2, px: 2, pb: 12 }}>
-      <Box sx={{ pt: 1, pb: 1, px: 1 }}>
-        <Typography level="h3" fontWeight="lg">Hi {firstName} 👋</Typography>
-        <Typography level="body-sm" color="neutral">Welcome Back</Typography>
-      </Box>
-      <Typography level="h3" sx={{ mb: 3, mt: 2 }}>My Rooms</Typography>
+      <Box sx={{ maxWidth: 480, mx: 'auto', mt: 2, px: 2, pb: 12 }}>
+        <Box sx={{ pt: 1, pb: 1, px: 1 }}>
+          <Typography level="h3" fontWeight="lg">Hi {firstName} 👋</Typography>
+          <Typography level="body-sm" color="neutral">Welcome Back</Typography>
+        </Box>
+        <Typography level="h3" sx={{ mb: 3, mt: 2 }}>My Rooms</Typography>
 
-      {rooms.length === 0 ? (
-        <Card variant="outlined" sx={{ textAlign: 'center', p: 4 }}>
-          <CardContent>
-            <Typography level="body-md" color="neutral">
-              You are not in any room yet.
-            </Typography>
-          </CardContent>
-        </Card>
-      ) : (
-        rooms.map(room => (
-          <Card
-            key={room.id}
-            variant="outlined"
-            sx={{
-              mb: 2,
-              cursor: 'pointer',
-              borderColor: 'primary.200',
-              boxShadow: 'sm',
-              transition: 'box-shadow 0.15s, border-color 0.15s',
-              '&:hover': { boxShadow: 'md', borderColor: 'primary.400' },
-              '&:active': { boxShadow: 'xs' },
-            }}
-            onClick={() => router.push(`/${room.id}`)}
-          >
+        {rooms.length === 0 ? (
+          <Card variant="outlined" sx={{ textAlign: 'center', p: 4 }}>
             <CardContent>
-              <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                <Box>
-                  <Typography level="title-lg" fontWeight="lg">Room #{room.id}</Typography>
-                  <Typography level="body-sm" color="neutral" sx={{ mt: 0.5 }}>
-                    {room.members} member{room.members !== 1 ? 's' : ''} · {room.role}
-                  </Typography>
-                </Box>
-                <Box sx={{
-                  bgcolor: 'primary.softBg',
-                  color: 'primary.plainColor',
-                  borderRadius: '999px',
-                  px: 2,
-                  py: 0.5,
-                  fontSize: 'sm',
-                  fontWeight: 'md',
-                }}>
-                  Open →
-                </Box>
-              </Box>
+              <Typography level="body-md" color="neutral">
+                You are not in any room yet.
+              </Typography>
             </CardContent>
           </Card>
-        ))
-      )}
-    </Box>
+        ) : (
+          rooms.map(room => (
+            <Card
+              key={room.id}
+              variant="outlined"
+              sx={{
+                mb: 2,
+                cursor: 'pointer',
+                borderColor: 'primary.200',
+                boxShadow: 'sm',
+                transition: 'box-shadow 0.15s, border-color 0.15s',
+                '&:hover': { boxShadow: 'md', borderColor: 'primary.400' },
+                '&:active': { boxShadow: 'xs' },
+              }}
+              onClick={() => router.push(`/${room.id}`)}
+            >
+              <CardContent>
+                <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+                  <Box>
+                    <Typography level="title-lg" fontWeight="lg">Room #{room.id}</Typography>
+                    <Typography level="body-sm" color="neutral" sx={{ mt: 0.5 }}>
+                      {room.members} member{room.members !== 1 ? 's' : ''} · {room.role}
+                    </Typography>
+                  </Box>
+                  <Box sx={{
+                    bgcolor: 'primary.softBg',
+                    color: 'primary.plainColor',
+                    borderRadius: '999px',
+                    px: 2,
+                    py: 0.5,
+                    fontSize: 'sm',
+                    fontWeight: 'md',
+                  }}>
+                    Open →
+                  </Box>
+                </Box>
+              </CardContent>
+            </Card>
+          ))
+        )}
+      </Box>
 
       <Box sx={{
         position: 'fixed',
